Rename Makefile -> Makefile.legacy
I want to add cmake configuration as it will help to turn
this project into a full-fledged library (and it is easier
to maintain and solve dependencies).
CMake generates `Makefile` which would overwrite the current
one, which I want to keep as a legacy file for now.

If cmake is not run, running `make -f Makefile.legacy` should still work
for some time.